Generic payload trailer support for PUT requests
This patch adds support for payload trailer that proxy server
sends to the object server at the end of object data stream
during a PUT request.  Useful when the object data is modified
at the proxy server (encrypted, erasure coded etc) before
streaming it down to the object server.  In such cases the
object server has no way of knowing/calcuating the original
object size (in the erasure coding case) and/or the md5sum for
the original object data (in the encryption/erasure coding
cases).  The proxy server needs to send this information to
the object server so the object server can store the object
metadata accurately along with the object and also in the
container database.

Changes:

 * Add a helper class for payload trailer support
   (swift/common/request_helpers.py:ObjectPayloadTrailer)
 * Proxy object controller modifications to insert object
   trailer at the end of PUT request payload
 * Object server modifications to extract object metadata
   and write to the object xattrs.  Note that the metadata
   sent as part of the payload trailer is opaque to the
   object server
 * Modifications to the object server GET method to
   return 'X-Object-ETag' metadata value for object etag
   and 'X-Object-Content-Length' value for object length.
 * Similar modifications to the object server HEAD method.
   This is so HEAD on the object server returns correct
   object length and etag even though the original object
   data has been altered.
 * Unit test code ObjectPayloadTrailer class

WIP - add trailer support to object unit test cases

Change-Id: I78b65b91debf2894cf020fe58834e423612b07c5
Implements: blueprint ec-proxy-work